Allow tests to be repeatable no matter the current state of the db
I drop the tables in the setup since to allow the tests to start from a clean slate since the state of your db from dev'ing is still there making the tests volatile. The clearing of the alembic_version table is to allow me to run the upgrade scripts again after running a test. If it's not cleared, the upgrade thinks all the upgrades have already run even though the tests just dropped all the tables.